+++ proc-macro-definitions/src/part/component.rs
@@ -143,8 +143,9 @@ impl<C> ParseWithContext for Component<C> {
 					parse2(quote_spanned! (open_span.resolved_at(Span::mixed_site())=> #path::render_args_builder())).expect("render_params make_builder 1"),
 					render_params.as_slice(),
 				);
-				parse2(quote_spanned! (open_span=> .render(bump, #render_params)))
-				.map_err(|_| Error::new(open_span, "Internal Asteracea error: Child component element didn't produce parseable capture"))?}
+				parse2(quote_spanned! (open_span=> .render(bump, #render_params)?))
+				.map_err(|_| Error::new(open_span, "Internal Asteracea error: Child component element didn't produce parseable attached access"))?
+			}
 		})
 		}
 	}
@@ -177,7 +178,7 @@ impl<C> Component<C> {
 				);
 				let mut expr = parse2(quote_spanned!(open_span.resolved_at(Span::mixed_site())=> {
 					#binding
-					reference.render(#bump, #render_params)
+					reference.render(#bump, #render_params)?
 				}))
 				.unwrap();
 				visit_expr_mut(&mut SelfMassager, &mut expr);