+++ tests/box.rs
@@ -1,6 +1,6 @@
 use std::pin::Pin;
 
-use asteracea::error::ExtractableResolutionError;
+use debugless_unwrap::DebuglessUnwrap;
 use lignin::bumpalo::Bump;
 use rhizome::Node;
 
@@ -29,16 +29,15 @@ asteracea::component! {
 }
 
 #[test]
-fn simple() -> Result<(), ExtractableResolutionError> {
+fn simple() {
 	let root = Node::new_for::<()>();
-	let component = Simple::new(&root.into(), Simple::new_args_builder().build())?;
+	let component =
+		Simple::new(&root.into(), Simple::new_args_builder().build()).debugless_unwrap();
 
 	let bump = Bump::new();
 	let _vdom = Box::pin(component)
 		.as_ref()
 		.render(&bump, Simple::render_args_builder().build());
-
-	Ok(())
 }
 
 asteracea::component! {
@@ -48,9 +47,10 @@ asteracea::component! {
 }
 
 #[test]
-fn named() -> Result<(), ExtractableResolutionError> {
+fn named() {
 	let root = Node::new_for::<()>();
-	let component = Box::pin(Named::new(&root.into(), Named::new_args_builder().build())?);
+	let component =
+		Box::pin(Named::new(&root.into(), Named::new_args_builder().build()).debugless_unwrap());
 
 	let bump = Bump::new();
 	let _vdom = component
@@ -58,8 +58,6 @@ fn named() -> Result<(), ExtractableResolutionError> {
 		.render(&bump, Named::render_args_builder().build());
 
 	let _: Boxed = component.named.boxed;
-
-	Ok(())
 }
 
 mod a_module {
@@ -75,14 +73,12 @@ mod a_module {
 }
 
 #[test]
-fn public() -> Result<(), ExtractableResolutionError> {
+fn public() {
 	use a_module::Public;
 
 	let root = Node::new_for::<()>();
-	let component = Box::pin(Public::new(
-		&root.into(),
-		Public::new_args_builder().build(),
-	)?);
+	let component =
+		Box::pin(Public::new(&root.into(), Public::new_args_builder().build()).debugless_unwrap());
 
 	let bump = Bump::new();
 	let _vdom = component
@@ -90,8 +86,6 @@ fn public() -> Result<(), ExtractableResolutionError> {
 		.render(&bump, Public::render_args_builder().build());
 
 	let _: a_module::Boxed = component.public.boxed;
-
-	Ok(())
 }
 
 asteracea::component! {
@@ -101,9 +95,10 @@ asteracea::component! {
 }
 
 #[test]
-fn typed() -> Result<(), ExtractableResolutionError> {
+fn typed() {
 	let root = Node::new_for::<()>();
-	let component = Box::pin(Typed::new(&root.into(), Typed::new_args_builder().build())?);
+	let component =
+		Box::pin(Typed::new(&root.into(), Typed::new_args_builder().build()).debugless_unwrap());
 
 	let bump = Bump::new();
 	let _vdom = component
@@ -112,8 +107,6 @@ fn typed() -> Result<(), ExtractableResolutionError> {
 
 	let typed: Pin<&TypedBoxed> = component.named.as_ref();
 	let _: Boxed = typed.boxed;
-
-	Ok(())
 }
 
 struct BoxContainer {
@@ -142,12 +135,11 @@ asteracea::component! {
 }
 
 #[test]
-fn reused() -> Result<(), ExtractableResolutionError> {
+fn reused() {
 	let root = Node::new_for::<()>();
-	let component = Box::pin(TypeReused::new(
-		&root.into(),
-		TypeReused::new_args_builder().build(),
-	)?);
+	let component = Box::pin(
+		TypeReused::new(&root.into(), TypeReused::new_args_builder().build()).debugless_unwrap(),
+	);
 
 	let bump = Bump::new();
 	let _vdom = component
@@ -156,8 +148,6 @@ fn reused() -> Result<(), ExtractableResolutionError> {
 
 	let typed: &Pin<Box<BoxContainer>> = &component.named;
 	let _: Boxed = typed.boxed;
-
-	Ok(())
 }
 
 asteracea::component! {
@@ -179,16 +169,14 @@ asteracea::component! {
 }
 
 #[test]
-fn multi() -> Result<(), ExtractableResolutionError> {
+fn multi() {
 	let root = Node::new_for::<()>();
-	let component = Multi::new(&root.into(), Multi::new_args_builder().build())?;
+	let component = Multi::new(&root.into(), Multi::new_args_builder().build()).debugless_unwrap();
 
 	let bump = Bump::new();
 	let _vdom = Box::pin(component)
 		.as_ref()
 		.render(&bump, Multi::render_args_builder().build());
-
-	Ok(())
 }
 
 asteracea::component! {