+++ tests/defer.rs
@@ -1,4 +1,3 @@
-use asteracea::error::ExtractableResolutionError;
 use lignin::bumpalo::Bump;
 use rhizome::Node;
 
@@ -16,14 +15,12 @@ use rhizome::Node;
 // }
 
 // #[test]
-// fn defer() -> Result<(), ExtractableResolutionError> {
+// fn defer() {
 // 	let root = Node::new_for::<()>();
-// 	let component = Deferrer::new(&root.into(), Deferrer::new_args_builder().build())?;
+// 	let component = Deferrer::new(&root.into(), Deferrer::new_args_builder().build());
 
 // 	let bump = Bump::new();
 // 	let _vdom = Box::pin(component)
 // 		.as_ref()
 // 		.render(&bump, Deferrer::render_args_builder().build());
-
-// 	Ok(())
 // }